refactor: Category 패키지 Catalog로 이름 변경
기존 Category enum은 이름 그대로 유지, 상위 개념 이름이 Catalog로만 된
것
