+++ src/main/java/kr/co/automl/domain/metadata/catalog/Catalog.java
@@ -1,11 +1,11 @@
-package kr.co.automl.domain.metadata.category;
+package kr.co.automl.domain.metadata.catalog;
 
-import kr.co.automl.domain.metadata.category.dto.CategoryDto;
+import kr.co.automl.domain.metadata.catalog.dto.CategoryDto;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 
 @EqualsAndHashCode
-public class CategoryEntity {
+public class Catalog {
 
     private long id;
 
@@ -14,24 +14,24 @@ public class CategoryEntity {
     private String themeTaxonomy;
 
     @Builder
-    private CategoryEntity(Category category, Theme theme, String themeTaxonomy) {
+    private Catalog(Category category, Theme theme, String themeTaxonomy) {
         this.category = category;
         this.theme = theme;
         this.themeTaxonomy = themeTaxonomy;
     }
 
-    public static CategoryEntity create(String categoryName, String themeName, String themeTaxonomy) {
+    public static Catalog create(String categoryName, String themeName, String themeTaxonomy) {
         Category category = Category.ofName(categoryName);
         Theme theme = category.findThemeByName(themeName);
 
-        return CategoryEntity.builder()
+        return Catalog.builder()
                 .category(category)
                 .theme(theme)
                 .themeTaxonomy(themeTaxonomy)
                 .build();
     }
 
-    public static CategoryEntity from(CategoryDto categoryDto) {
+    public static Catalog from(CategoryDto categoryDto) {
         return create(
                 categoryDto.name(),
                 categoryDto.theme(),
