+++ src/main/java/kr/co/automl/domain/metadata/catalog/Category.java
@@ -1,7 +1,7 @@
-package kr.co.automl.domain.metadata.category;
+package kr.co.automl.domain.metadata.catalog;
 
-import kr.co.automl.domain.metadata.category.exceptions.CannotFindMatchCategoryException;
-import kr.co.automl.domain.metadata.category.exceptions.CannotFindMatchThemeException;
+import kr.co.automl.domain.metadata.catalog.exceptions.CannotFindMatchCategoryException;
+import kr.co.automl.domain.metadata.catalog.exceptions.CannotFindMatchThemeException;
 
 import java.util.Arrays;
 import java.util.Objects;
