+++ src/main/java/kr/co/automl/domain/metadata/catalog/dto/CategoryDto.java
@@ -1,4 +1,4 @@
-package kr.co.automl.domain.metadata.category.dto;
+package kr.co.automl.domain.metadata.catalog.dto;
 
 import lombok.Builder;
