+++ src/test/java/kr/co/automl/domain/metadata/catalog/CatalogTest.java
@@ -1,16 +1,16 @@
-package kr.co.automl.domain.metadata.category;
+package kr.co.automl.domain.metadata.catalog;
 
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class CategoryEntityTest {
+class CatalogTest {
 
     @Test
     void create_생성_테스트() {
-        CategoryEntity categoryEntity = CategoryEntity.create("대기 환경", "공기질", "themeTaxonomy");
+        Catalog catalog = Catalog.create("대기 환경", "공기질", "themeTaxonomy");
 
-        assertThat(categoryEntity).isEqualTo(CategoryEntity.builder()
+        assertThat(catalog).isEqualTo(Catalog.builder()
                 .category(Category.ATMOSPHERIC_ENVIRONMENT)
                 .theme(Theme.AIR_QUALITY)
                 .themeTaxonomy("themeTaxonomy")
